+++ components/layout/sequential.rs
@@ -6,7 +6,7 @@
 
 use app_units::Au;
 use context::LayoutContext;
-use display_list_builder::{DisplayListBuildState, StackingContextCollectionState};
+use display_list::{DisplayListBuildState, StackingContextCollectionState};
 use euclid::{Point2D, Vector2D};
 use floats::SpeculatedFloatPlacement;
 use flow::{Flow, ImmutableFlowUtils, FlowFlags, GetBaseFlow};
